fix broken autoconf cache variable zebra_rtread
It looks like someone tried to make this cached, but it was being
ignored at least on autoconf 2.6.1.  2.6.2 now gives a
warning about the situation, hence this fix.

Although they are just warnings, it's dangerous to leave them
as-is, because someone not diligent in reading the autoconf manual
might just stick _cv_ in the name, with the effect that running
a ./configure --config-cache would fail horribly, since variables
are being set inside the AC_CACHE_CHECK() block that would not
be seen when the variable's value was read from cache.

Also added m4 quotes [] to macro arguments as autoconf manual
suggests.