Introduce lib/symtab.c & .h
These are two brand new files.

They provide a "Symbol Table" data structure, which maps symbol "names"
to symbol "values.  (Inside the dta structure is a hash table.)

There is also support for managing references to symbols, and the
ability to visit all references to a symbol when its value is
set or unset.

Other files affected:

  lib/Makefile.am

    -- to add the new files.

  lib/memtypes.c

    -- additional memory types for symbol table stuff