Allow orphan symbols to be unset -- lib/symtab.c
May not set orphan symbol values to anything meaningful, but must be
allowed to unset an orphan value (which is already unset) so that
when reaming a symbol table ordinary value delete/unset functions can
be called.

So fixed the assert which was disallowing all value changes for
orphans, in symbol_set_value().