Further refinement to arguments for message queue blocks.
Updates lib/mqueue.c and .h

Allows for main message argument to be a (small) embedded structure.

Users of messages can now define suitable structs for the information
to be passed in a message, and use those to write/read the
message data.  Should be easier to use and less error prone than
trying to match anonymous argv[n] values.

Allows for straightforward, flexible argv[] array as well.