Binding to interfaces and counting of messages.
Wired up message counters in bgp_session structure.

Added fields to session for neighbor interface and
neighbor update-source -- so that these can be set when connect()
is done.  Peering Engine resolves any interface name to an address,
so that BGP Engine doesn't have to.

Reinstated as much code as necessary in bgp_network to bind to
specific interfaces, as set in the session.

Moved setting of bgp_nexthop_set() back into Routeing Engine.

Result is that only Peering Engine talks to Zebra or uses the
iflist.

Wired up setting of TTL.

Reworked connections locking of the session mutex so more robust
if/when connections are cut loose from the session.

Made peer_index entry point at connection, not session.  Works
better in bgp_network that way.

  modified:   bgpd/bgp_connection.c
  modified:   bgpd/bgp_connection.h
  modified:   bgpd/bgp_fsm.c
  modified:   bgpd/bgp_msg_read.c
  modified:   bgpd/bgp_msg_write.c
  modified:   bgpd/bgp_network.c
  modified:   bgpd/bgp_network.h
  modified:   bgpd/bgp_peer.c
  modified:   bgpd/bgp_peer.h
  modified:   bgpd/bgp_peer_index.c
  modified:   bgpd/bgp_peer_index.h
  modified:   bgpd/bgp_session.c
  modified:   bgpd/bgp_session.h
  modified:   lib/prefix.h
  modified:   lib/sockunion.c
  modified:   lib/sockunion.h