Added test for sEstablished before processing incoming updates.
Before disabling a session the Peering Engine now revokes its inbound
queue for the session.  This clears out anything that may have been
queued up to that moment for the session.

After disabling a session, some updates may still arrive, because
it may be a while before the BGP Engine actions the disable, and
stops queueing stuff.  So, Peering Engine must check that the session
is established before processing stuff in the queue.

  modified:   bgpd/bgp_engine.h
  modified:   bgpd/bgp_session.c