Put back some Quagga additions to getopt.c etc.
Also, backed out a small change in zebra/misc_null.c, which removed
a warning, but on reflection may be using deep magic which is not
fully understood (by me, anyway).  zebra appears to have a number
of definitions for the same function, and how that is all linked
is not obvious (to me, anyway).