Fixing declarations to eliminate compiler warnings in zebra.
Only parts of zebra/*.c are compiled and linked, depending on what
was chosen at "configure" time.  A subset of that is compiled and
linked for testzebra.

Some things were not declared, or not declared everywhere they
were required... leading to a number of compiler warnings.

These changes are intended to tidy that up.