Bring up to date with 0.99.16 release
This brings this branch up to date with 0.99.16, EXCEPT for the
change to bgpd to use monotonic clock for timing updates.

There are places where bgpd expects to find the actual time that
things happened... so this change is not though to be complete.

Further tidying in zebra, for FreeBSD.