Don't return "Not Supported" when setting NULL MD5 password
If TCP MD5 signature is not supported, sockopt_tcp_signature()
returns ENOSYS.

This change means that if the given password is NULL or an empty
string, then returns OK, even when TCP MD5 signature is not
supported.

If the password is being cleared, it doesn't matter that it
could never have bee set in the first place !

This means that unless a password is required, it is safe to call
sockopt_tcp_signature() whether the facility is supported or not.
Which further means that only this function needs to worry about
the circumstances under which it is supported !