Fixes to vty stuff so that zebra runs.
Changes to vty handling require the library to behave one way
when running as part of the new (threaded) bgpd and another way
when running as part of any of the other Quagga daemons.

Code to revoke things from the command queue failed to test
whether it was running as part of the new bgpd or not.