Introduce "rs-in" route-map
All peers may be configured with an "rs-in" route-map, whether they
are rsclients, or not.

All routes from a given peer are passed through the "rs-in" route-map
before being considered for inclusion in any rsclients RIBs.

Routes pass through the "rs-in" route-map once, irrespective of the
number of rsclients.

An "rs-in" route-map is intended to be used to ensure that routes
from a given peer are kosher, before being passed to rsclients.

The "in" route-map and filters may do the same, but for the main RIB.
Any route-selection type actions that the "in" route-map may do for
the main RIB will probably be done in the "import" route-map for an
rsclient RIB.

Also:

  * make zfree() work for NULL items, in lib/memory.c

  * fix clear ip bgp .... so works when pEstablished

Version change to ex03.