Avoid creating neighbor with same address as another.
With bgp multiple-instance it was possible to create a neighbor in one
"view" with the same addres as a neighbor in another view (or the main,
unnamed view).

To do so was a mistake, since (in particular) the accept() mechanism
depends on finding the neighbor by its address -- so two neighbors with
the same address == confusion.

Attempting to create two neighbors with the same address is now
treated as an error, at the command level.

Also, added "-rdynamic" to the CFLAGS for gcc in configure.ac.  This
improves the output from the stack traceback in the event of an
assert or other failure.