Fix bug in NOTIFICATION handling before FSM reaches Established state.
Change removes the setting of session->active to false which was being
done before a NOTIFICATION message is sent.  This flag should remain
true when the session is not being stopped -- which is the case if
a NOTIFICATION is sent before Established state is reached.

Effect of this bug was to trip up the accept() side of the session,
bringing bgpd down on an assert().

This bug may be triggered by a peer that accepts a connection and then
remains silent, for whatever reason -- causing bgpd to issue an
HoldTimer Expired NOTIFICATION.

Version advanced to 0.99.15ex07.

Other changes purely cosmetic -- eg changing some 'int' to 'bool',
and a few small documentation edits.