Initial commit to seed the "pipework" branch
This is a major revision of the command processing, in order to support
new lexical level for command lines, plus all the necessary I/O
redirection for the pipes.

This is version 0.99.15ex10p.

This supports:

  < filename
  <+ filename
  ....   > filename
  ....   >> filename
  ....   >*

Also contains all lexical level handling of '...', "...." and \x in order
to allow use of '>' et al if required.

Updated command line completion and help is a work in progress.