Improve writing of pid lock file and error reporting.
Added fsync() after writing the pid to the lock file to push the result
to the file system immediately.

Any error in the procedure is now treated as fatal, and the daemon will
exit(1).  (Previously only the open and the lock had to succeed.)

Any error is logged and is (now) output to stderr.

If the pid file is locked, the error message reports the pid which owns
the lock.