Updated the "symbol table" and added AVL tree
Modified the symbol table so that the name of the symbol is now
a property of its value -- which simplifies some of the semantics,
because the name no longer has a separate existence (a name which
refers to an empty or undefined value is now a matter for the
value, not for the symbol table).  It also allows for the symbol
table to be used where the value and the name are almost
indistinguishable.

Introduced AVL tree to support large prefix lists, routemaps,
etc. etc.

Updated testsymtab and added testavl.

Updated exiting use of the symbol table.