vtysh partly restored.  bgpd not crashing on over-size messages.
NB: this is a partial interim version -- NOT for production.

Reinstate vtysh.  In bgpd: cope better with generating oversize messages.

The "pipework" branch is no more.  There is now "euro_ix" and "euro_ix_b",
where 'b' stands for 'bleedin'.  The "pipework" is currently in 'bleedin'.

Version updated to: 0.99.18ex20b

Major changes in this commit:

* modified bgpd to avoid crashing if an oversize message is created.

  At present, any oversize messages are simply discarded.  This may be a
  mistake... since it is possible that some routes that should have been
  withdrawn are not.  TBA.

  The stream lib facilities have been overhauled to may this easier.

* restoring vtysh.

  At present the vtysh is thought to work, except that it does not
  currently create integrated configuration files.

For the time being, *only* the following compile:

  lib, bgpd, zebra, vtysh, tests

All other daemons will fail to compile, and must be disabled.