Fix broken prefix-list handling in bgpd
Some changes to symtab.c had broken the symbol reference handling for
prefix-lists in plist.c

This commit further "improves" the symbol table functions, to support
current use rather better.  One key change is that the symbol table
supports the notion of a symbol which is "set", so has some value other
than its name.

In the case of prefix-lists, a prefix-list is "set" when there is at
least one (valid) entry or a description has been set.

Changes include:

  * small changes in bgpd/bgp_clist.c where symbol table function names
    have changed and other small differences.

  * similarly in bgpd/bgp_peer_index.c, but here also clarified ownership
    of the peer index entries, and ensured that the symbol table stuff
    does *not* free same.

  * in bgpd/bgp_route.c fixed the handling of references to prefix-lists.
    Newer symbol handling more clearly identifies when a symbol has some
    value "set".

  * in bgpd/bgp_vty.c similarly, when showing state of filtering the "*"
    against a prefix-list means it is "set".

  * in lib/plist.c -- adjust to newer symbol table stuff.  Fix issues
    will add_hook and delete_hook.  Changed lookup so that if plist
    is not set, then lookup will return not-found -- so works as
    before (!).

  * adjusted test for symbol stuff.