Handle IPv6 as pairs of uint64_t in prefix.c.
C99 requires uint64_t, so seems reasonable to use those where possible.

Reduces most IPv6 functions to operations on one or two 64-bit parts.

For native 32-bit processors, we leave it to the compiler to optimise the
operation on pairs of 32-bit values -- and look forward to such processors
being phased out.

For maximum performance requires a __builtin_bswap64() or equivalant to be
added.