Fix compilation issue with <features.h>, which is a glibc special.
The "misc.h" #includes <features.h> because that was found to give consistent
values for the many macros that control GNU, many flavours of POSIX, and
Standard C/C99 library behaviour -- for glibc.

Sadly, this broke compilation where <features.h> is absent, ie when glibc is
not used.

To fix this made changes to configure.ac to create HAVE_FEATURES_H, so that
the #include can be skipped.

At the same time, added AC_FUNC_STRERROR_R and removed the *horrible* kludge
which to date has supported use of strerror_r.

Tidied up and re-tested use of strerror_r().