Remove CONFIRM(_POSIX_REENTRANT_FUNCTIONS) -- not POSIX !
Turns out that _POSIX_REENTRANT_FUNCTIONS is not a POSIX feature
test macro... name notwithstanding.

It appears to be quite well known, but in some cases appears to be a
feature request, and in others a feature test :-(  So, since it adds
no value -- and breaks FreeBSD -- the CONFIRM is removed.

Further work in progress for Watch-Dog -- all known mutexes are held
in an all known mutexes list, complete with name of mutex for diagnostic
purposes.