ripd: pad simple password properly
RFC2453 4.1 defines RIP-2 simple password data unit as follows: "The
remaining 16 octets contain the plain text password. If the password is
under 16 octets, it must be left-justified and padded to the right with
nulls (0x00)."

The function checking simple password in received packets,
rip_auth_check_password(), used strncmp() instead of memcmp(). This
would lead to an incorrectly (a null followed by non-nulls) padded
password to be accepted as a correct one.

The function writing a simple password to the packet authentication
header, rip_auth_write_leading_rte(), always used 16 bytes of a given
string, regardless of the actual size allocated and the value of any
bytes beyond the terminating null.