zebra: use SA_SIZE for RT_ROUNDUP on FreeBSD
FreeBSD provides SA_SIZE (and none of the other options to infer padded
size of a struct sockaddr).  Just define SAROUNDUP to SA_SIZE if it is
available.

This also drops a superfluous-looking extra macro branch which would
require ROUNDUP.  It seemed redundant to my eyes, but I have no idea
what odd things might have triggered addition of this in the first
place...

Signed-off-by: David Lamparter <equinox@opensourcerouting.org>