zebra-nexthop-tracking.patch
Added support for nexthop tracking of a prefix.

Support for protocols to register zebra to notify them when a nexthop becomes
valid. With this support, protocols such as BGP no longer have to periodically
scan the RIB checking if a prefix is valid or invalid. Complete documentation
in doc/next-hop-tracking.txt.

[TT: Rebased on top of master.]

Signed-off-by: Pradosh Mohapatra <pmohapat@cumulusnetworks.com>
Signed-off-by: Daniel Walton <dwalton@cumulusnetworks.com>
Signed-off-by: Dinesh Dutt <ddutt@cumulusnetworks.com>