lib: wrong #define used for IPV6_MINHOPCOUNT
The #define IPV6_MINHOPCNT define is never defined on any unix platform.
>From what I can tell the original implementation on the linux platform
was IPV6_MINHOPCNT, when it got accepted into the mainstream kernel
it was transformed into IPV6_MINHOPCOUNT.  Since we test for the
define before attempting to use the code it was silently doing nothing
for a long time.

Signed-off-by: David Lamparter <equinox@opensourcerouting.org>