bgpd: improve ebgp-multihop and ttl-security handling
Allow setting ebgp-multihop for iBGP. It will basically only set
TTL for iBGP connections as the other features of ebgp-multihop
affect only eBGP. This improves compatibility with other
implementations that allow setting TTL for all connections.
Enabling this has similar reasoning as for enabling gtsm for ibgp
(see commit 5d804b439a4138c77).

Change also to track configured value in ->ttl and ->gtsm_hops;
not the value set to sockopt. Instead, setting of socket's ttl
and minttl options are now merged to one function which calculates
it on demand. This greatly simplifies the code.

Signed-off-by: Timo Teräs <timo.teras@iki.fi>

Conflicts:
	bgpd/bgpd.c