Add support for run quagga in a FIB (Routing table) on FreeBSD
Patch to add support for FreeBSD's fib. if quagga is running in a fib,
quagga will only trick to grab interfaces which are set to be in the
same FIB as the quagga process.

When gathering the system routing table, only grab the one corresponding
to the fib the process is running in.

https://bugzilla.quagga.net/show_bug.cgi?id=839