nhrpd: detect ipsec binding updates
Track IPsec SA unique ids better, and detect when the address
binding changes. This purges cache information for those peers,
and additionally tear down IPsec SA for cache entries with
registration, to force the client to re-register.

If and when we do FlexVPN style IKEv2 routing, these hooks can
be used directly to update the NHRP cache with the updated address.