vtysh: Allow display of individual daemons configs
When executing a 'show run' or 'write terminal' you see the
entire integrated config.  You have no way of knowing what an
individual daemon is going to write until after you do a write
of config to disk if you are not using an integrated configuration.
This change allows the end-user to do such a thing.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>