bgpd: Correctly install non LL route into zebra
If a route being installed into the kernel from bgp is a
ipv6 global address, set the ifindex to 0 and let the kernel
figure out the correct interface to use.

This fixes the bug where a third-party nexthop is received for
an alternate interface than the incoming bgp message is being
rejected by the linux kernel because the ifindex chosen is for
the interface of the received packet not the interface of the
routes nexthop.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>